+++ internal/crossdock/thrift/gauntlet/secondserviceclient/client.go
@@ -50,10 +50,12 @@ type Interface interface {
 //
 // 	client := secondserviceclient.New(dispatcher.ClientConfig("secondservice"))
 func New(c transport.ClientConfig, opts ...thrift.ClientOption) Interface {
-	return client{c: thrift.New(thrift.Config{
-		Service:      "SecondService",
-		ClientConfig: c,
-	}, opts...)}
+	return client{
+		c: thrift.New(thrift.Config{
+			Service:      "SecondService",
+			ClientConfig: c,
+		}, opts...),
+	}
 }
 
 func init() {
@@ -62,7 +64,9 @@ func init() {
 	})
 }
 
-type client struct{ c thrift.Client }
+type client struct {
+	c thrift.Client
+}
 
 func (c client) BlahBlah(
 	ctx context.Context,