+++ internal/crossdock/thrift/gauntlet/thrifttestclient/client.go
@@ -170,10 +170,12 @@ type Interface interface {
 //
 // 	client := thrifttestclient.New(dispatcher.ClientConfig("thrifttest"))
 func New(c transport.ClientConfig, opts ...thrift.ClientOption) Interface {
-	return client{c: thrift.New(thrift.Config{
-		Service:      "ThriftTest",
-		ClientConfig: c,
-	}, opts...)}
+	return client{
+		c: thrift.New(thrift.Config{
+			Service:      "ThriftTest",
+			ClientConfig: c,
+		}, opts...),
+	}
 }
 
 func init() {
@@ -182,7 +184,9 @@ func init() {
 	})
 }
 
-type client struct{ c thrift.Client }
+type client struct {
+	c thrift.Client
+}
 
 func (c client) TestBinary(
 	ctx context.Context,