+++ internal/examples/thrift-hello/hello/echo/helloclient/client.go
@@ -45,10 +45,12 @@ type Interface interface {
 //
 // 	client := helloclient.New(dispatcher.ClientConfig("hello"))
 func New(c transport.ClientConfig, opts ...thrift.ClientOption) Interface {
-	return client{c: thrift.New(thrift.Config{
-		Service:      "Hello",
-		ClientConfig: c,
-	}, opts...)}
+	return client{
+		c: thrift.New(thrift.Config{
+			Service:      "Hello",
+			ClientConfig: c,
+		}, opts...),
+	}
 }
 
 func init() {
@@ -57,7 +59,9 @@ func init() {
 	})
 }
 
-type client struct{ c thrift.Client }
+type client struct {
+	c thrift.Client
+}
 
 func (c client) Echo(
 	ctx context.Context,