+++ scripts/generate.sh
@@ -39,6 +39,9 @@ thriftrw --plugin=yarpc --out=internal/examples/thrift-keyvalue/keyvalue interna
 thriftrw --out=encoding/thrift encoding/thrift/internal.thrift
 thriftrw --out=serialize serialize/internal.thrift
 
+thriftrw --no-recurse --plugin=yarpc --out=encoding/thrift/thriftrw-plugin-yarpc/testdata encoding/thrift/thriftrw-plugin-yarpc/testdata/common.thrift
+thriftrw --no-recurse --plugin=yarpc --out=encoding/thrift/thriftrw-plugin-yarpc/testdata encoding/thrift/thriftrw-plugin-yarpc/testdata/atomic.thrift
+
 thrift-gen --generateThrift --outputDir internal/crossdock/thrift/gen-go --inputFile internal/crossdock/thrift/echo.thrift
 thrift-gen --generateThrift --outputDir internal/crossdock/thrift/gen-go --inputFile internal/crossdock/thrift/gauntlet_tchannel.thrift